events in lessons
Now there is support for events in lessons. I'm not really happy with
the design yet, but at least it works.

For now the only supported event is triggered when something gets
executed in the console.

It should now be possible to quite easily port the try ruby tutorial
to the new lessons architecture.