Simple patch that will make sure that selections can only be made by clicking and dragging within the text.
The old behavior allowed residual dragging from other widgets to start selecting the text in the text field. Also caused by a small issue where the release methods are not fired when the mouse is released over another window or alert. This is mitigated by basically augmenting the logic of the mouse[0] field with another instance variable that denotes a click has been made within the widget.

Some cleanup of hard tabs to spaces was done in other parts by coincidence. Discipline!

Closes #41